changes done for multi module adding modules jdbc updated

+++ demo/customer-service-standalone/src/main/resources/schema.sql
@@ -1,6 +1,10 @@
-drop table customer;
+drop table customer if exists;
 
 create table customer (
-    id number not null,
-    customer_name varchar2(25)
+    id number not null AUTO_INCREMENT,
+    customer_name varchar2(25),
+    PRIMARY KEY (id)
+
 );
+
+insert into customer(customer_name)  values ('gopal');
